fix: preserve authored markdown in flavor exports
