+++ tests/_convert/markdown/test_markdown_from_ir.py
@@ -8,11 +8,9 @@ from marimo._convert.markdown.flavor import (
 )
 from marimo._convert.markdown.flavor.base import (
     CodeCellBlock,
-    DirectiveBlock,
     MarkdownCellBlock,
     MarkdownExportDocument,
 )
-from marimo._convert.markdown.flavor.pymdown import split_pymdown_blocks
 from marimo._convert.markdown.from_ir import (
     _format_filename_title,
     _get_sql_options_from_cell,
@@ -425,303 +423,6 @@ def test_qmd_flavor_preserves_explicit_filters():
     )
 
 
-def test_qmd_flavor_maps_pymdown_admonitions_to_callouts():
-    """Test that qmd flavor renders PyMdown admonitions as Quarto callouts."""
-    flavor = markdown_flavor_from_filename("notebook.qmd")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """Before
-
-/// attention | Careful
-
-This needs attention.
-///
-
-After"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert '::: {.callout-important title="Careful"}' in markdown
-    assert "This needs attention.\n:::" in markdown
-    assert "Before\n\n::: {.callout-important" in markdown
-    assert ":::\n\nAfter" in markdown
-
-
-def test_qmd_flavor_preserves_plain_pymdown_callout_title():
-    """Test that plain PyMdown callout titles stay unquoted in QMD."""
-    flavor = markdown_flavor_from_filename("notebook.qmd")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// tip | Variables panel
-
-Open the variables panel.
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert '::: {.callout-tip title="Variables panel"}' in markdown
-
-
-def test_qmd_flavor_preserves_quoted_pymdown_callout_title():
-    """Test that quoted PyMdown callout titles stay quoted in QMD."""
-    flavor = markdown_flavor_from_filename("notebook.qmd")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// tip | "Variables panel"
-
-Open the variables panel.
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert '::: {.callout-tip title="&quot;Variables panel&quot;"}' in markdown
-
-
-def test_qmd_flavor_maps_generic_admonition_type_to_callout():
-    """Test that generic PyMdown admonition type selects Quarto callout type."""
-    flavor = markdown_flavor_from_filename("notebook.qmd")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// admonition | Heads up
-    type: warning
-
-Watch this.
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert '::: {.callout-warning title="Heads up"}' in markdown
-    assert "Watch this.\n:::" in markdown
-
-
-def test_pymdown_flavor_preserves_pymdown_admonitions():
-    """Test that pymdown flavor keeps PyMdown syntax unchanged."""
-    flavor = markdown_flavor_from_filename("notebook.md")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// tip | Keep this
-
-PyMdown syntax is preserved.
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert "/// tip | Keep this" in markdown
-
-
-def test_convert_from_ir_to_markdown_maps_admonitions_for_qmd():
-    """Test full export maps PyMdown markdown admonitions for qmd output."""
-    app = App()
-
-    @app.cell()
-    def __():
-        import marimo as mo
-
-        return (mo,)
-
-    @app.cell()
-    def __(mo):
-        mo.md(
-            """
-            /// tip | Tip with Title
-
-            This is an example.
-            ///
-            """
-        )
-        return
-
-    internal_app = InternalApp(app)
-    notebook = internal_app.to_ir()
-
-    markdown = convert_from_ir_to_markdown(notebook, filename="notebook.qmd")
-
-    assert '::: {.callout-tip title="Tip with Title"}' in markdown
-    assert "This is an example.\n:::" in markdown
-
-
-def test_qmd_flavor_maps_pymdown_tabs_to_panel_tabsets():
-    """Test that consecutive PyMdown tabs become a Quarto panel tabset."""
-    flavor = markdown_flavor_from_filename("notebook.qmd")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// tab | Python
-print("py")
-///
-
-/// tab | SQL
-    select: true
-
-select 1
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert "::: {.panel-tabset}" in markdown
-    assert '## Python\n\nprint("py")' in markdown
-    assert "## SQL {.active}\n\nselect 1" in markdown
-
-
-def test_qmd_flavor_starts_new_tabset_for_pymdown_new_tab_option():
-    """Test that PyMdown tab new option starts another Quarto tabset."""
-    flavor = markdown_flavor_from_filename("notebook.qmd")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// tab | A
-A
-///
-
-/// tab | B
-    new: true
-
-B
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert markdown.count("::: {.panel-tabset}") == 2
-    assert "## A\n\nA" in markdown
-    assert "## B\n\nB" in markdown
-
-
-def test_qmd_flavor_falls_back_to_pandoc_divs():
-    """Test that unmapped PyMdown directives become Quarto-compatible divs."""
-    flavor = markdown_flavor_from_filename("notebook.qmd")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// details | More
-    attrs: {id: more, class: folded quiet}
-    open: true
-
-Body
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert (
-        '::: {.details title="More" #more .folded .quiet open="true"}'
-        in markdown
-    )
-    assert "Body\n:::" in markdown
-
-
-def test_split_pymdown_blocks_keeps_directives_inside_markdown_fences():
-    """Test that literal directives inside code fences are not parsed."""
-    blocks = split_pymdown_blocks(
-        """```text
-literal
-``` not a closing fence
-/// tip | Should stay code
-body
-///
-```
-
-/// tip | Real
-body
-///"""
-    )
-
-    assert len(blocks) == 2
-    assert isinstance(blocks[0], MarkdownCellBlock)
-    assert "/// tip | Should stay code" in blocks[0].text
-    assert isinstance(blocks[1], DirectiveBlock)
-    assert blocks[1].argument == "Real"
-
-
-def test_split_pymdown_blocks_preserves_body_colon_lines():
-    """Test that body-leading colon lines are not consumed as options."""
-    blocks = split_pymdown_blocks(
-        """/// details | Example
-    key: value
-    still body
-///"""
-    )
-
-    assert len(blocks) == 1
-    assert isinstance(blocks[0], DirectiveBlock)
-    assert blocks[0].options == {}
-    assert blocks[0].body == "    key: value\n    still body"
-
-
-def test_mystmd_flavor_maps_pymdown_blocks_to_myst_directives():
-    """Test that mystmd flavor maps PyMdown blocks to MyST directives."""
-    flavor = normalize_markdown_flavor("mystmd", filename="notebook.md")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// tip | Nice
-Body
-///
-
-/// tab | Python
-print("py")
-///
-
-/// tab | SQL
-select 1
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert ":::{tip} Nice\nBody\n:::" in markdown
-    assert "::::{tab-set}" in markdown
-    assert ':::{tab-item} Python\nprint("py")\n:::' in markdown
-    assert ":::{tab-item} SQL\nselect 1\n:::" in markdown
-
-
 def test_mystmd_flavor_renders_marimo_notebook_export_syntax():
     """Test that mystmd flavor renders marimo notebook authoring syntax."""
     flavor = normalize_markdown_flavor("mystmd", filename="notebook.md")
@@ -787,28 +488,3 @@ def test_mystmd_flavor_grows_code_fence_guard():
 
     assert "````{marimo} python" in markdown
     assert markdown.rstrip().endswith("````")
-
-
-def test_mystmd_flavor_merges_classes_into_one_option():
-    """Test that MyST directives do not repeat the class option."""
-    flavor = normalize_markdown_flavor("mystmd", filename="notebook.md")
-    document = MarkdownExportDocument(
-        metadata={"title": "Notebook"},
-        header=None,
-        blocks=[
-            MarkdownCellBlock(
-                """/// admonition | Heads up
-    type: tip
-    attrs: {class: extra}
-
-Body
-///"""
-            )
-        ],
-    )
-
-    markdown = flavor.render_document(document)
-
-    assert (
-        ":::{admonition} Heads up\n:class: tip extra\n\nBody\n:::" in markdown
-    )
